requirements.txt
# AWS Strands - AIエージェント開発フレームワーク
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
strands-agents-builder>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# テストフレームワーク
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# PDF生成
reportlab>=4.0.0

# Excel生成
openpyxl>=3.1.0

# 画像処理（image_readerツール用）
Pillow>=10.0.0

# データ処理
python-dateutil>=2.8.2

# 環境変数管理
python-dotenv>=1.0.0

# ロギング（標準ライブラリだが明示的に記載）
# logging - 標準ライブラリ
